角色架构悠

+++ src/views/AddressBook/comp/role/rolelist.vue
@@ -26,10 +26,10 @@
               color="indigo">mdi-home</v-icon>
           </v-list-tile-action>
           <v-list-tile-content
-            @click="handleSelect(item.title)">
+            @click="handleSelect(item)">
             <v-list-tile-title>{{ item.title }}</v-list-tile-title>
           </v-list-tile-content>
-          <v-list-tile-action>
+          <!-- <v-list-tile-action>
             <v-menu
               bottom
               left
@@ -56,9 +56,9 @@
                 </v-list>
               </v-card>
             </v-menu>
-          </v-list-tile-action>
+          </v-list-tile-action> -->
           <v-list-tile-action
-            @click="handleSelect(item.title)">
+            @click="handleSelect(item)">
             <v-icon>mdi-arrow-right-thick</v-icon>
           </v-list-tile-action>
         </v-list-tile>
@@ -75,49 +75,26 @@
               large
               color="indigo">mdi-account</v-icon>
           </v-list-tile-action>
-
-          <v-list-tile-content>
+          <v-list-tile-content @click="deleteUser(item)">
             <v-list-tile-title>{{ item.username }}</v-list-tile-title>
           </v-list-tile-content>
-          <v-list-tile-action>
-            <v-menu
-              bottom
-              left
-              content-class="dropdown-menu"
-              offset-y
-              transition="slide-y-transition">
-              <div
-                v-ripple
-                slot="activator"
-              >
-                <v-icon>mdi-settings</v-icon>
-              </div>
-              <v-card>
-                <v-list dense>
-                  <v-list-tile
-                    v-for="route in userroute"
-                    :key="route"
-                    @click="set(route, item)"
-                  >
-                    <v-list-tile-title
-                      v-text="route"
-                    />
-                  </v-list-tile>
-                </v-list>
-              </v-card>
-            </v-menu>
-          </v-list-tile-action>
           <v-list-tile-action>
             <v-icon :color="item.active ? 'teal' : 'grey'">mdi-account</v-icon>
           </v-list-tile-action>
         </v-list-tile>
       </v-list>
     </v-card>
+    <!-- ---------------------------菜单----------------------->
+    <Menu :role="currentRole"/>
   </v-flex>
 </template>
 <script>
 import { findAllRoletreeAstree, findRoles, deleteRoleById } from '@/api/roletree'
+import Menu from './Menu'
 export default {
+  components: {
+    Menu
+  },
   props: {
     selecteditems: {
       type: Array,
@@ -138,14 +115,8 @@ export default {
       // 部门
       items2: [],
       selected1: {},
-      noderoute: [
-        '删除节点',
-        '添加角色',
-        '角色添加成员'
-      ],
-      userroute: [
-        '删除成员'
-      ]
+      // 当前选中的角色
+      currentRole: {}
     }
   },
   watch: {
@@ -158,11 +129,8 @@ export default {
     },
     selected1: {
       handler (newval, oldval) {
-        // console.log(newval)
-        if (newval.type === '角色') {
-          // 员工更新
-          this.findUserWithRole(newval.title)
-        }
+        this.currentRole = newval
+        this.findUserWithRole(newval.title)
         this.$emit('update:selected', this.selected) // 当前选中的部门信息
       }
     }
@@ -171,10 +139,9 @@ export default {
     this.getDatas()
   },
   methods: {
-    handleSelect (role) {
-      // router 更新
-      // console.log(department)
-      this.selecteditems.push(role)
+    handleSelect (item) {
+      this.selecteditems.push(item.title)
+      this.currentRole = item
       this.$emit('update:selecteditems', this.selecteditems)
     },
     // 设置当前选中角色
@@ -214,6 +181,7 @@ export default {
       findAllRoletreeAstree({ company: this.company }).then(res => {
         this.data = res.data
         this.items2 = this.data[0].children
+        this.currentRole = this.data[0]
         this.selecteditems.push(this.data[0].title)
       })
     },
@@ -226,62 +194,15 @@ export default {
         this.items = res.data
       })
     },
-    set (route, item) {
-      switch (route) {
-        case '添加角色':
-          this.addRole(item)
-          break
-        case '角色添加成员':
-          this.addUser(item)
-          break
-        case '删除成员':
-          this.deleteUser(item)
-          break
-        default:
-          this.$Message.info('暂无此功能')
-      }
-    },
     deleteUser (item) {
-      deleteRoleById(item.id).then(res => {
-        this.$Message.info(res.data.message)
-      })
-    },
-    addUser (item) {
-      if (!this.permission()) {
-        this.$Message.error('需要【OA管理员】角色')
-        return
-      }
-      if (item.type !== '角色') {
-        this.$Message.error('只有【角色】才能添加成员')
-        return
-      }
-      this.$router.push({
-        name: '角色添加成员',
-        query: {
-          role: item.title
-        }
-      })
-    },
-    addRole (item) {
-      if (!this.permission()) {
-        this.$Message.error('需要【OA管理员】角色')
-        return
-      }
       // console.log(item)
-      if (item.type !== '节点') {
-        this.$Message.error('只有【节点】才能添加角色')
-        return
-      }
-      this.$router.push({
-        name: '添加角色',
-        query: {
-          role: item
-        }
-      })
-    },
-    permission () {
-      return this.$store.state.user.roles.some(role => {
-        return role === 'OA管理员'
+      this.$Confirm.confirm('确定要删除：' + item.username).then(() => {
+        deleteRoleById(item.id).then(res => {
+          this.$Message.info(res.data.message)
+        })
+        // console.log('确定删除')
+      }).catch(() => {
+        // console.log('不删除')
       })
     }
   }
